+++ examples/data-sources/meraki_sensor_relationships/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_sensor_relationships" "example" {
+  serial = "1234-ABCD-1234"
+}