+++ examples/resources/meraki_sensor_relationships/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_sensor_relationships.example "<serial>"