+++ examples/resources/meraki_sensor_relationships/resource.tf
@@ -1,0 +1,8 @@
+resource "meraki_sensor_relationships" "example" {
+  serial = "1234-ABCD-1234"
+  livestream_related_devices = [
+    {
+      serial = ""
+    }
+  ]
+}